+++ pcap4j-core/src/test/java/org/pcap4j/core/PcapHandleTest.java
@@ -1,13 +1,20 @@
 package org.pcap4j.core;
 
 import static org.junit.Assert.*;
+
+import java.util.ArrayList;
 import java.util.List;
+
 import org.junit.After;
 import org.junit.AfterClass;
 import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.Test;
+import org.pcap4j.core.BpfProgram.BpfCompileMode;
+import org.pcap4j.core.PcapHandle.PcapDirection;
 import org.pcap4j.core.PcapNetworkInterface.PromiscuousMode;
+import org.pcap4j.packet.IcmpV4EchoPacket;
+import org.pcap4j.packet.IcmpV4EchoReplyPacket;
 import org.pcap4j.packet.Packet;
 import org.pcap4j.packet.namednumber.DataLinkType;
 
@@ -178,6 +185,76 @@ public class PcapHandleTest {
         assertEquals(packet.length, ph.getOriginalLength().intValue());
       }
     });
+  }
+
+  @Test
+  public void testSetDirection() throws Exception {
+    if (System.getenv("TRAVIS") != null) {
+      // run only on Travis CI
+      PcapNetworkInterface nif = Pcaps.getDevByName("any");
+      PcapHandle handle = nif.openLive(65536, PromiscuousMode.PROMISCUOUS, 10);
+      handle.setDirection(PcapDirection.OUT);
+      handle.setFilter("icmp", BpfCompileMode.OPTIMIZE);
+
+      ProcessBuilder pb = new ProcessBuilder("/bin/ping", "www.google.com");
+      Process process = pb.start();
+
+      final List<Packet> packets = new ArrayList<Packet>();
+      handle.loop(3, new PacketListener() {
+        @Override
+        public void gotPacket(Packet packet) {
+          packets.add(packet);
+        }
+      });
+      handle.close();
+      process.destroy();
+
+      assertEquals(3, packets.size());
+      assertTrue(packets.get(0).contains(IcmpV4EchoPacket.class));
+      assertTrue(packets.get(1).contains(IcmpV4EchoPacket.class));
+      assertTrue(packets.get(2).contains(IcmpV4EchoPacket.class));
+    }
+    else {
+      try {
+        ph.setDirection(PcapDirection.OUT);
+        fail();
+      } catch (PcapNativeException e) {
+        assertTrue(e.getMessage().startsWith("Failed to set direction:"));
+      }
+    }
+  }
+
+  @Test
+  public void testDirection() throws Exception {
+    if (System.getenv("TRAVIS") != null) {
+      // run only on Travis CI
+      PcapHandle handle
+        = new PcapHandle.Builder("any")
+            .direction(PcapDirection.IN)
+            .promiscuousMode(PromiscuousMode.PROMISCUOUS)
+            .snaplen(65536)
+            .timeoutMillis(10)
+            .build();
+      handle.setFilter("icmp", BpfCompileMode.OPTIMIZE);
+
+      ProcessBuilder pb = new ProcessBuilder("/bin/ping", "www.google.com");
+      Process process = pb.start();
+
+      final List<Packet> packets = new ArrayList<Packet>();
+      handle.loop(3, new PacketListener() {
+        @Override
+        public void gotPacket(Packet packet) {
+          packets.add(packet);
+        }
+      });
+      handle.close();
+      process.destroy();
+
+      assertEquals(3, packets.size());
+      assertTrue(packets.get(0).contains(IcmpV4EchoReplyPacket.class));
+      assertTrue(packets.get(1).contains(IcmpV4EchoReplyPacket.class));
+      assertTrue(packets.get(2).contains(IcmpV4EchoReplyPacket.class));
+    }
   }
 
 }